blux_modules/security/contain_respond/process_isolator.py
#!/usr/bin/env python3
"""
BLUX Guard — Process isolator
- Snapshot process state & allow rollback
- Quarantine suspicious processes
"""
import os
import logging
import psutil  # Requires: pip install psutil

# Configure logging
logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
logger = logging.getLogger(__name__)

def snapshot_process(pid):
    """Snapshots process state (memory, open files, network connections)."""
    try:
        process = psutil.Process(pid)
        snapshot = {
            "pid": pid,
            "name": process.name(),
            "exe": process.exe(),
            "cwd": process.cwd(),
            "status": process.status(),
            "create_time": process.create_time(),
            "open_files": process.open_files(),
            "connections": process.connections(),
            "memory_info": process.memory_full_info()._asdict(),
        }
        # Create a filename based on the pid to hold the snapshot.
        snapshot_filename = f"/tmp/process_snapshot_{pid}.json"
        # Ensure file path
        os.makedirs(os.path.dirname(snapshot_filename), exist_ok = True)
        with open(snapshot_filename, "w") as f:
            import json
            json.dump(snapshot, f, indent=4)
        logger.info(f"[ISOLATE] PID={pid} snapshot saved to {snapshot_filename}")
        return snapshot_filename

    except psutil.NoSuchProcess:
        logger.error(f"[ISOLATE] PID={pid} not found.")
        return None
    except Exception as e:
        logger.exception(f"[ISOLATE] Error snapshotting PID={pid}: {e}")
        return None

def isolate_process(pid):
    """Quarantines a process."""
    # Placeholder: implement actual quarantine (cgroups, namespaces, etc.)
    logger.warning(f"[ISOLATE] PID={pid} quarantined.  (No quarantine is currently implemented)")
    # Placeholder: Implement rollback logic
    return True

def rollback_process(pid, snapshot_file):
    """Rolls back a process to a previous state using a snapshot."""
    if not os.path.exists(snapshot_file):
        logger.error(f"[ISOLATE] Snapshot does not exist {snapshot_file}")
        return False
    # Placeholder: Implement actual rollback logic
    logger.warning(f"[ISOLATE] PID={pid} rollback: (No rollback is currently implemented)")
    return True

if __name__ == "__main__":
    # Test Isolation
    test_pid = os.getpid()
    logger.info(f"Running test, snapshotting PID: {test_pid}")
    snapshot_file = snapshot_process(test_pid)
    if snapshot_file:
        logger.info(f"Test rollback to: {snapshot_file}")
        rollback_process(test_pid, snapshot_file)
        isolate_process(test_pid)
